% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{da_vacinacao_municipio}
\alias{da_vacinacao_municipio}
\title{da_vacinacao_municipio}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 77325 rows and 10 columns.
}
\usage{
da_vacinacao_municipio
}
\description{
da_vacinacao_municipio
}
\keyword{datasets}
